+++ crates/wdl-lsp/src/server.rs
@@ -209,7 +209,7 @@ impl ProgressToken {
 }
 
 /// Represents options for running the LSP server.
-#[derive(Debug, Default)]
+#[derive(Debug, Default, Clone)]
 pub struct ServerOptions {
     /// The name of the server.
     ///
@@ -241,42 +241,44 @@ pub struct Server {
 }
 
 impl Server {
+    /// Creates a new WDL language server.
+    pub fn new(client: Client, options: ServerOptions) -> Self {
+        let lint = options.lint;
+        let analyzer_client = client.clone();
+        Self {
+            client,
+            options,
+            analyzer: Analyzer::<ProgressToken>::new_with_validator(
+                DiagnosticsConfig::new(rules()),
+                move |token, kind, current, total| {
+                    let client = analyzer_client.clone();
+                    async move {
+                        let message = format!(
+                            "{kind} {current}/{total} file{s}",
+                            s = if total > 1 { "s" } else { "" }
+                        );
+                        let percentage = ((current * 100) as f64 / total as f64) as u32;
+                        token.update(&client, message, percentage).await
+                    }
+                },
+                move || {
+                    let mut validator = Validator::default();
+                    if lint {
+                        validator.add_visitor(Linter::default());
+                    }
+                    validator
+                },
+            ),
+            client_support: Default::default(),
+            folders: Default::default(),
+        }
+    }
+
     /// Runs the server until a request is received to shut down.
     pub async fn run(options: ServerOptions) -> Result<()> {
         debug!("running LSP server: {options:#?}");
 
-        let (service, socket) = LspService::new(|client| {
-            let lint = options.lint;
-            let analyzer_client = client.clone();
-
-            Self {
-                client,
-                options,
-                analyzer: Analyzer::<ProgressToken>::new_with_validator(
-                    DiagnosticsConfig::new(rules()),
-                    move |token, kind, current, total| {
-                        let client = analyzer_client.clone();
-                        async move {
-                            let message = format!(
-                                "{kind} {current}/{total} file{s}",
-                                s = if total > 1 { "s" } else { "" }
-                            );
-                            let percentage = ((current * 100) as f64 / total as f64) as u32;
-                            token.update(&client, message, percentage).await
-                        }
-                    },
-                    move || {
-                        let mut validator = Validator::default();
-                        if lint {
-                            validator.add_visitor(Linter::default());
-                        }
-                        validator
-                    },
-                ),
-                client_support: Default::default(),
-                folders: Default::default(),
-            }
-        });
+        let (service, socket) = LspService::new(|client| Self::new(client, options.clone()));
 
         let stdin = tokio::io::stdin();
         let stdout = tokio::io::stdout();